Drop unused CSS selectors.
This roughly cuts the CSS size in half.

Implemented as a script because it has to run post build to be able to see all generated HTML (including from markdown). Changed the CSS hashing to include our own CSS code as well as the names and versions of the module dependencies that deal with css.